requirements.txt
django==5.1.7
pillow==11.1.0
django-crispy-forms==2.3
django-taggit==6.1.0
django-ckeditor==6.7.2
whitenoise==6.9.0
gunicorn==23.0.0 